# cmd/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:9044",
    "basePath": "/",
    "paths": {
...
        "/api/start-external-login": {
            "post": {
                "description": "starts an external login ceremony with an external IDP.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "starts an external login ceremony with an external IDP",
                "parameters": [
                    {
                        "description": "StartExternalIDPLoginRequest",
                        "name": "external_idp",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/external_idp.StartExternalIDPLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/external_idp.StartExternalIDPLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/external-idp": {
            "post": {
                "description": "externalIDP.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "root"
                ],
                "summary": "todo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "internalCode": {
                    "type": "string"
                }
            }
        },
...
        "external_idp.StartExternalIDPLoginRequest": {
            "type": "object",
            "required": [
                "directive",
                "slug"
            ],
            "properties": {
                "directive": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "external_idp.StartExternalIDPLoginResponse": {
            "type": "object",
            "properties": {
                "redirectUri": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}